requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-CORS
Flask-Bcrypt
psycopg2-binary
gunicorn
python-dotenv
